feat(linker): try to push category catching in refresolver but hit a wall
